+++ vscode-client/extension.ts
@@ -2,6 +2,7 @@ import * as vscode from "vscode";
 import { GherkinFormattingEditProvider } from "./formattingEditProvider";
 import { GherkinDefinitionProvider } from "./definitionProvider";
 import { GherkinHoverProvider } from "./hoverProvider";
+import { GherkinReferenceProvider } from "./referenceProvider";
 
 export async function activateAsync(context: vscode.ExtensionContext): Promise<void> {
   const robotcode = vscode.extensions.getExtension("d-biehl.robotcode");
@@ -16,6 +17,7 @@ export async function activateAsync(context: vscode.ExtensionContext): Promise<v
 
   const definitionProvider = new GherkinDefinitionProvider();
   const hoverProvider = new GherkinHoverProvider();
+  const referenceProvider = new GherkinReferenceProvider();
 
   context.subscriptions.push(
     vscode.languages.registerDocumentFormattingEditProvider("gherkin", new GherkinFormattingEditProvider()),
@@ -23,6 +25,7 @@ export async function activateAsync(context: vscode.ExtensionContext): Promise<v
     vscode.languages.registerDefinitionProvider("markdown", definitionProvider),
     vscode.languages.registerHoverProvider("gherkin", hoverProvider),
     vscode.languages.registerHoverProvider("markdown", hoverProvider),
+    vscode.languages.registerReferenceProvider("robotframework", referenceProvider),
   );
 }